Move commons to cloud client base class
